refactor(tests): migrate notifications route tests to real PostgreSQL
Replace DB mock with real test DB for notification_log queries.
Fix COUNT(*)::integer cast in route for correct number type from pg.

+++ backend/src/routes/notifications.ts
@@ -37,7 +37,7 @@ export async function notificationRoutes(fastify: FastifyInstance) {
     `, [...params, limit, offset]);
 
     const total = await db.queryOne<{ count: number }>(`
-      SELECT COUNT(*) as count FROM notification_log ${where}
+      SELECT COUNT(*)::integer as count FROM notification_log ${where}
     `, [...params]);
 
     return {
